inititial starting point (0.7)

+++ php/tor_client.php
@@ -6,12 +6,17 @@
 
 
 function transmission_sessionId() {
-  global $config_values;
+  global $config_values, $platform;
   $sessionIdFile = get_tr_sessionIdFile();
+  if(file_exists($sessionIdFile) && !is_writable($sessionIdFile)) {
+      $myuid = posix_getuid();
+      echo "<div id=\"errorDialog\" class=\"dialog_window\" style=\"display: block\">$sessionIdFile is not writable for uid: $myuid</div>";
+      return;
+  }
 
   if(file_exists($sessionIdFile)) {
       if(filesize($sessionIdFile) > 0) {
-        $handle = fopen($sessionIdFile, r);
+        $handle = fopen($sessionIdFile, 'r');
         $sessionId = trim(fread($handle, filesize($sessionIdFile)));
     } else {
         unlink($sessionIdFile);
@@ -30,7 +35,7 @@ function transmission_sessionId() {
                     CURLOPT_NOBODY => true,
                     CURLOPT_USERPWD => "$tr_user:$tr_pass"
                   );
-    get_curl_defaults(&$curl_options);
+    get_curl_defaults($curl_options);
 
     curl_setopt_array($sid, $curl_options);
 
@@ -39,19 +44,26 @@ function transmission_sessionId() {
   
     preg_match("/X-Transmission-Session-Id:\s(\w+)/",$header,$ID);
 
-    $handle = fopen($sessionIdFile, "w");
-    fwrite($handle, $ID[1]);
-    fclose($handle);
-  
-    $sessionId = $ID[1];
+    if(isset($ID[1])) {
+      $handle = fopen($sessionIdFile, "w");
+      fwrite($handle, $ID[1]);
+      fclose($handle);
+      if($platform == 'NMT') chmod($sessionIdFile, 0666);
+      $sessionId = $ID[1];
+    }
   }
-
-  return $sessionId;
+  if(isset($sessionId))
+    return $sessionId;
 }
 
 function transmission_rpc($request) {
   global $config_values;
   $sessionIdFile = get_tr_sessionIdFile();  
+  if(file_exists($sessionIdFile) && !is_writable($sessionIdFile)) {
+      $myuid = posix_getuid();
+      echo "<div id=\"errorDialog\" class=\"dialog_window\" style=\"display: block\">$sessionIdFile is not writable for uid: $myuid</div>";
+      return;
+  }
 
   $tr_user = $config_values['Settings']['Transmission Login'];
   $tr_pass = get_client_passwd();
@@ -79,13 +91,12 @@ function transmission_rpc($request) {
                                                ),
                           CURLOPT_POSTFIELDS => "$request"
                        );
-    get_curl_defaults(&$curl_options);
+    get_curl_defaults($curl_options);
     curl_setopt_array($post, $curl_options);
 
     $raw = curl_exec($post);
     curl_close($post);
-
-    if(preg_match('/409: Conflict/', $raw)) {
+    if(preg_match('/409:? Conflict/', $raw)) {
         if(file_exists($sessionIdFile)) {
             unlink($sessionIdFile);
         }
@@ -99,17 +110,30 @@ function transmission_rpc($request) {
 function get_deep_dir($dest, $tor_name) {
     global $config_values;
     switch($config_values['Settings']['Deep Directories']) {
-      case '0':
+    case '0':
         break;
-      case 'Title':
+    case 'Title_Season':
+          $guess = guess_match($tor_name, TRUE);
+          if(isset($guess['key']) && isset($guess['episode'])) {
+            if(preg_match('/^(\d{1,3})x\d+p?$/', $guess['episode'], $Season)) {
+                $dest = $dest."/".ucwords(strtolower($guess['key']))."/Season ".$Season[1];
+            } else if(preg_match('/^(\d{4})\d{4}$/', $guess['episode'], $Year)) {
+                $dest = $dest."/".ucwords(strtolower($guess['key']))."/".$Year[1];
+            } else {
+                $dest = $dest."/".ucwords(strtolower($guess['key']));
+            }
+            break;
+          }
+          _debug("Deep Directories: Couldn't match $tor_name Reverting to Full\n", 1);
+    case 'Title':
         $guess = guess_match($tor_name, TRUE);
         if(isset($guess['key'])) {
           $dest = $dest."/".ucwords(strtolower($guess['key']));
           break;
         }
         _debug("Deep Directories: Couldn't match $tor_name Reverting to Full\n", 1);
-      case 'Full':
-      default:
+    case 'Full':
+    default:
         $dest = $dest."/".ucwords(strtolower($tor_name));
         break;
     }
@@ -138,8 +162,6 @@ function transmission_add_torrent($tor, $dest, $title, $seedRatio) {
                                        )
                                );
   $response = transmission_rpc($request);
-  _debug(json_encode($response),0);
-  _debug("\r\n",0);
 
   $torHash = $response['arguments']['torrent-added']['hashString'];
 
@@ -150,12 +172,10 @@ function transmission_add_torrent($tor, $dest, $title, $seedRatio) {
              'seedRatioMode' => 1)
             );
     $response = transmission_rpc($request);
-    _debug(json_encode($response),0);
-    _debug("\r\n",0);
   } 
 
 
-  if(isset($response['result']) AND ($response['result'] == 'success' or $response['result'] == 'duplicate torrent')) {
+  if(isset($response['result']) AND ($response['result'] == 'success')) {
     $cache = $config_values['Settings']['Cache Dir'] . "/rss_dl_" . filename_encode($title);
     if($torHash) {
       $handle = fopen("$cache", "w");
@@ -163,6 +183,8 @@ function transmission_add_torrent($tor, $dest, $title, $seedRatio) {
       fclose($handle);
     }
     return 0;
+  } else if ($response['result'] == 'duplicate torrent') {
+      return "Duplicate Torrent";
   } else {
     if(!isset($response['result']))
       return "Failure connecting to Transmission";
@@ -171,8 +193,9 @@ function transmission_add_torrent($tor, $dest, $title, $seedRatio) {
   }
 }
 
-function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL) {
+function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL, $retried=false) {
   global $config_values, $hit;
+  if(strtolower($fav['Filter']) == "any") $any=1;
   $hit = 1;
   $filename = htmlspecialchars_decode($filename);
 
@@ -181,7 +204,6 @@ function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL
   if($feed && preg_match('/:COOKIE:/', $feed) && (!(preg_match('/:COOKIE:/', $url)))) {
     $url .= stristr($feed, ':COOKIE:');    
   }
-  
   $get = curl_init();
   $response = check_for_cookies($url);
   if($response) {
@@ -189,17 +211,29 @@ function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL
       $cookies = $response['cookies'];
   }
   $getOptions[CURLOPT_URL] = $url;
-  $getOptions[CURLOPT_COOKIE] = $cookies;
+  if(isset($cookies)) $getOptions[CURLOPT_COOKIE] = $cookies;
   $getOptions[CURLOPT_USERAGENT] = 'Python-urllib/1.17';  
-  get_curl_defaults(&$getOptions);
+  $getOptions[CURLOPT_FOLLOWLOCATION] = true;
+  get_curl_defaults($getOptions);
   curl_setopt_array($get, $getOptions);
   $tor = curl_exec($get);
   curl_close($get);
-
-  if(!($tor)) {
+  if (strncasecmp($tor, 'd8:announce', 11) != 0) { // Check for torrent magic-entry
+	//This was not a torrent-file, so it's poroperly some kind og xml / html.
+	if(!$retried) {
+	    //Try to retrieve a .torrent link from the content.
+	    $link = find_torrent_link($url, $tor);
+	    return client_add_torrent($link, $dest, $title, $feed, $fav, $url);
+	} else {
+	    _debug("No torrent file found on $url. Exitting.\n");
+	    if(isset($retried)) $url = $retried;
+	    return "Error: No torrent file found on $url.";
+	}
+  }
+  if(!$tor) {
   print '<pre>'.print_r($_GET, TRUE).'</pre>';
     _debug("Couldn't open torrent: $filename \n",-1);
-    return FALSE;
+    return "Error: Couldn't open torrent: $filename";
   }
   
   $tor_info = new BDecode("", $tor);
@@ -212,10 +246,10 @@ function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL
   }
   if(isset($fav) && $fav['Save In'] != 'Default') {
     $dest = $fav['Save In'];
-  } else if($config_values['Settings']['Deep Directories']) {
-    $dest = get_deep_dir($dest, $tor_name);
-    _debug("Deep Directorys, change dest to $dest\n", 1);
   }
+
+  $dest = get_deep_dir($dest, $tor_name);
+
   if(!file_exists($dest) or !is_dir($dest)) {
     $old_umask = umask(0);
     if(file_exists($dest))
@@ -249,29 +283,78 @@ function client_add_torrent($filename, $dest, $title, $feed = NULL, &$fav = NULL
     add_history($tor_name);
     _debug("Started: $tor_name in $dest\n",0);
     if(isset($fav)) {
-      run_script('favstart', $tor_name);
+      run_script('favstart', $title);
       if($config_values['Settings']['Email Notifications'] == 1) {
-          $subject = "TorrentWatch-X: Started downloading $tor_name";
+          $subject = "TW-X: $tor_name started downloading.";
           $msg = "TorrentWatch started downloading $tor_name";
           sendmail($msg, $subject);
       }
-      updateFavoriteEpisode($fav, $tor_name);
-      _debug("Updated Favorites");
+      if(!$any) {
+	updateFavoriteEpisode($fav, $title);
+        _debug("Updated Favorites");
+      }
     } else {
-        run_script('nonfavstart', $tor_name);
+        run_script('nonfavstart', $title);
     }
     if($config_values['Settings']['Save Torrents'])
       file_put_contents("$dest/$tor_name.torrent", $tor);
+    return "Success";
   } else {
     _debug("Failed Starting: $tor_name  Error: $return\n",-1);
 
     $msg = "TorrentWatch-X tried to start \"$tor_name\". But this failed with the following error:\n\n";
     $msg.= "$return\n";
 
-    $subject = "TorrentWatch-X: Error while trying to start $tor_name.";
+    $subject = "TW-X: Error while trying to start $tor_name.";
     sendmail($msg, $subject);
     run_script('error', $title, $msg);
+    return "Error: $return";
   }
-  return ($return === 0);
 }
+
+
+function find_torrent_link($url_old, $content) {
+	$url = "";
+	if($ret = preg_match('/["\']([^\'"]*?\.torrent[^\'"]*?)["\']/', $content, $matches)) {
+	    if (isset($ret)) {
+		$url = $matches[1];
+		if (!preg_match('/^https?:\/\//', $url)) {
+			if (preg_match('^/', $url)) {
+				$url = dirname($url_old) . $url;
+			} else {
+				$url = dirname($url_old) . '/' . $url;
+			}
+		}
+	    }
+	} else  {
+	    $ret = preg_match_all('/href=["\']([^#].+?)["\']/', $content, $matches);
+	    if ($ret) {
+		foreach($matches[1] as $match) {
+		    if (!preg_match('/^https?:\/\//', $match)) {
+			if (preg_match('^/', $match)) {
+			    $match = dirname($url_old) . $match;
+			} else {
+			    $match = dirname($url_old) . '/' . $match;
+			}
+		    }
+		    if (preg_match('/w3.org/i', $match)) {
+			break;
+		    }
+		    $opts = array('http' =>
+			array('timeout'=>10)
+		    );
+		    stream_context_get_default($opts);
+		    $headers = get_headers($match, 1);
+		    if((isset($headers['Content-Disposition']) && 
+		      preg_match('/filename=.+\.torrent/i', $headers['Content-Disposition'])) ||
+		      (isset($headers['Content-Type']) &&
+		      $headers['Content-Type'] == 'application/x-bittorrent' )) {
+			    $url = $match;
+		    }
+		}
+	    }
+	}
+	return $url;
+}
+
 ?>
